feat(ui-commons): use DownloadButton in DownloadMatrixButton
